requirements.txt
langchain==0.1.0
langchain-community==0.0.13
langchain-google-genai==0.0.6
google-generativeai==0.3.2
chromadb==0.4.22
sentence-transformers==2.3.1
# requirements.txt additions
chromadb==0.4.22
sentence-transformers==2.3.1
openai==1.51.0
httpx==0.27.2
langchain==0.1.0
langchain-community==0.0.13
python-dotenv==1.0.0
tiktoken==0.5.2